feat: protected routes for each role
